utils: create .axes module
The sherpa.utils.axes module now contains the Axis class from the
sherpa.models.regrid module.

Type statements have been added.

Downstream code should be updated to access the PointAxis and
IntegratedAxis classes directly from sherpa.utils.axes rather than
sherpa.models.regrid.
