+++ sherpa/data.py
@@ -1,5 +1,5 @@
 #
-#  Copyright (C) 2008, 2015 - 2017, 2019 - 2025
+#  Copyright (C) 2008, 2015-2017, 2019-2025
 #  Smithsonian Astrophysical Observatory
 #
 #
@@ -51,8 +51,8 @@ The design for the `Data` class assumes
 
 - fields are converted to `ndarray` when read in.
 
-- the independent axes can either be points (`~sherpa.models.regrid.PointAxis`)
-  or integrated (`~sherpa.models.regrid.IntegratedAxis`). There is
+- the independent axes can either be points (`~sherpa.utils.axes.PointAxis`)
+  or integrated (`~sherpa.utils.axes.IntegratedAxis`). There is
   currently no support for an object with a combination of point and
   integrated axes but it could be added.
 
@@ -124,10 +124,11 @@ import warnings
 
 import numpy as np
 
-from sherpa.models.regrid import EvaluationSpace1D, IntegratedAxis, PointAxis
+from sherpa.models.regrid import EvaluationSpace1D
 from sherpa.utils import NoNewAttributesAfterInit, formatting, \
     print_fields, create_expr, create_expr_integrated, \
     calc_total_error, bool_cast, filter_bins
+from sherpa.utils.axes import IntegratedAxis, PointAxis
 from sherpa.utils.err import DataErr
 from sherpa.utils.numeric_types import SherpaFloat
 from sherpa.utils.parallel import parallel_map_funcs
