+++ sherpa/models/regrid.py
@@ -28,7 +28,6 @@ for signal outside the data range - and then be regridded to
 match the desired grid.
 """
 
-from abc import ABCMeta, abstractmethod
 import itertools
 import logging
 import warnings
@@ -38,7 +37,13 @@ from sherpa.utils._utils import rebin  # type: ignore
 from sherpa.utils.akima import akima
 
 from sherpa.astro.utils import reshape_2d_arrays
-from sherpa.utils.err import DataErr, ModelErr
+
+# Axis is imported in case any code tries to access it. This is a
+# temporary fix, as users should directly import from
+# sherpa.utils.axes instead.
+#
+from sherpa.utils.axes import Axis, PointAxis, IntegratedAxis
+from sherpa.utils.err import ModelErr
 
 warning = logging.getLogger(__name__).warning
 
@@ -46,235 +51,7 @@ warning = logging.getLogger(__name__).warning
 PIXEL_RATIO_THRESHOLD = 0.1
 
 
-def _to_readable_array(x):
-    """Convert x into a ndarray that can not be edited (or is None)."""
-
-    if x is None:
-        return None
-
-    x = np.asarray(x).copy()
-    if not np.iterable(x):
-        raise DataErr("notanarray")
-
-    if x.ndim != 1:
-        raise DataErr("not1darray")
-
-    x.setflags(write=False)
-    return x
-
-
-class Axis(metaclass=ABCMeta):
-    """Represent an axis of a N-D object."""
-
-    # This is set when the data is set
-    _is_ascending = None
-
-    @property
-    def is_empty(self):
-        """Is the axis empty?
-
-        An empty axis is either set to `None` or a zero-element
-        sequence.
-        """
-        return self.size == 0
-
-    @property
-    @abstractmethod
-    def is_integrated(self):
-        """Is the axis integrated?"""
-        pass
-
-    @property
-    def is_ascending(self):
-        """Is the axis ascending?
-
-        The axis is ascending if the elements in `lo` are sorted in
-        ascending order.  Only the first and last elements are
-        checked, and it is assumed that the elements are sorted.
-        """
-        if self.is_empty:
-            raise DataErr("Axis is empty or has a size of 0")
-
-        return self._is_ascending
-
-    @property
-    @abstractmethod
-    def start(self):
-        """The starting point (lowest value) of the data axis."""
-        pass
-
-    @property
-    @abstractmethod
-    def end(self):
-        """The ending point of the data axis
-
-        If the data axis is ascending the end boundary is the last
-        element of the `hi` array when the axis is integrated,
-        otherwise it's the last element of `lo`. Conversely, for
-        descending axes, the last element is either the first element
-        of the `hi` array or of the `lo` array, depending on whether
-        the axis is integrated or not, respectively.
-        """
-        pass
-
-    @property
-    @abstractmethod
-    def size(self):
-        """The size of the axis."""
-        pass
-
-    def overlaps(self, other):
-        """Check if this axis overlaps with another.
-
-        Parameters
-        ----------
-        other : Axis instance
-            The axis to compare to.
-
-        Returns
-        -------
-        bool
-            True if they overlap, False if not
-        """
-        # Could apply this check but this is not expected to be
-        # called directly, so leave for now.
-        #
-        # if not isinstance(other, Axis):
-        #     raise TypeError("other argument must be an axis")
-
-        num = max(0, min(self.end, other.end) - max(self.start, other.start))
-        return bool(num != 0)
-
-
-class PointAxis(Axis):
-    """Represent a point (not integrated) axis of a N-D object.
-
-    The length can not be changed once set.
-
-    Parameters
-    ----------
-    x : array_like or None
-        The starting point of the axis. If `None` or `[]` then the
-        data axis is said to be empty. The axis can be in ascending or
-        descending order but this is not checked.
-
-    """
-
-    def __init__(self, x):
-        self._x = _to_readable_array(x)
-        if self._x is None:
-            return
-
-        nx = len(self._x)
-        if nx > 0:
-            self._is_ascending = self._x[-1] > self._x[0]
-
-    @property
-    def x(self):
-        return self._x
-
-    @property
-    def is_integrated(self):
-        return False
-
-    @property
-    def start(self):
-        if self.is_ascending:
-            return self.x[0]
-
-        return self.x[-1]
-
-    @property
-    def end(self):
-        if self.is_ascending:
-            return self.x[-1]
-
-        return self.x[0]
-
-    @property
-    def size(self):
-        if self.x is None:
-            return 0
-
-        return self.x.size
-
-
-class IntegratedAxis(Axis):
-    """Represent an integrated axis of a N-D object.
-
-    Parameters
-    ----------
-    lo : array_like or None
-        The starting point of the axis.  If `lo` is `None` or `[]`
-        then the data axis is said to be empty. The axis can be in
-        ascending or descending order.
-    hi : array_like or None
-        The ending point of the axis. The number of elements must match `lo` (either `None`
-        or a sequence of the same size). Each element is expected to
-        be larger than the corresponding element of the `lo` axis,
-        even if the `lo` array is in descending order.
-
-    """
-
-    _lo = None
-    _hi = None
-
-    def __init__(self, lo, hi):
-        self._lo = _to_readable_array(lo)
-        self._hi = _to_readable_array(hi)
-
-        if self._lo is None:
-            if self._hi is None:
-                return
-
-            raise DataErr("mismatchn", "lo", "hi", "None", len(self._hi))
-
-        nlo = len(self._lo)
-        if self._hi is None:
-            raise DataErr("mismatchn", "lo", "hi", nlo, "None")
-
-        nhi = len(self._hi)
-        if nlo != nhi:
-            raise DataErr("mismatchn", "lo", "hi", nlo, nhi)
-
-        if nlo > 0:
-            self._is_ascending = lo[-1] > lo[0]
-
-    @property
-    def lo(self):
-        return self._lo
-
-    @property
-    def hi(self):
-        return self._hi
-
-    @property
-    def is_integrated(self):
-        return True
-
-    @property
-    def start(self):
-        if self.is_ascending:
-            return self.lo[0]
-
-        return self.lo[-1]
-
-    @property
-    def end(self):
-        if self.is_ascending:
-            return self.hi[-1]
-
-        return self.hi[0]
-
-    @property
-    def size(self):
-        if self.lo is None:
-            return 0
-
-        return self.lo.size
-
-
-class EvaluationSpace1D():
+class EvaluationSpace1D:
     """Class for 1D Evaluation Spaces.
 
     An Evaluation Space is a set of data axes representing the data
@@ -407,7 +184,7 @@ class EvaluationSpace1D():
         return self.start <= other.start and self.end >= other.end
 
 
-class EvaluationSpace2D():
+class EvaluationSpace2D:
     """Class for 2D Evaluation Spaces.
 
     An Evaluation Space is a set of data axes representing the data
@@ -559,7 +336,7 @@ class EvaluationSpace2D():
         return np.zeros(size)
 
 
-class ModelDomainRegridder1D():
+class ModelDomainRegridder1D:
     """Allow 1D models to be evaluated on a different grid.
 
     This class is not used directly in a model expression;
@@ -844,7 +621,7 @@ class ModelDomainRegridder1D():
                                         **kwargs)
 
 
-class ModelDomainRegridder2D():
+class ModelDomainRegridder2D:
     """Allow 2D models to be evaluated on a different grid.
 
     This class is not used directly in a model expression;
